thm2isis now uses out attrs and converted to func
