+++ isis/src/odyssey/apps/thm2isis/thm2isis.h
@@ -1,0 +1,12 @@
+#ifndef thm2isis_h // Change this to your app name in all lower case suffixed with _h (e.g. campt_h, cam2map_h etc.)
+#define thm2isis_h
+
+#include "Cube.h"
+#include "UserInterface.h"
+
+namespace Isis {
+
+  extern void thm2isis(UserInterface &ui);
+}
+
+#endif
